+++ ChangeLog.md
@@ -30,6 +30,7 @@
 * Prise en compte correcte des malus d'armure et de casques pour les compétences sur la fiche.
 
 ### Autres améliorations
+* Possibilité de faire des jets d'attaque contre des valeurs de caractéristiques
 * Prise en compte d'un prédicat attaqueSournoise pour indiquer le nombre de dés des attaques sournoises.
 * Si une arme est aussi une arme de jet, le script va automatiquement sélectionner la bonne attaque selon la distance de la cible.
 * Ajout d'une option --saveDM pour éviter tous les dégâts d'une attaque.